requirements.txt
# ============================================================
# PyTorch 版依赖（AutoDL: PyTorch 2.1.0 + CUDA 11.8 镜像）
# 注意：mmcv 必须用 mim 或预编译 wheel 安装，不能直接 pip install
# 完整安装步骤见 README.md 或 install.sh
# ============================================================

# MMDetection 生态（版本需严格对应 PyTorch/CUDA）
mmengine>=0.8.0,<1.0.0
mmdet>=3.2.0,<4.0.0
# mmcv 请通过 install.sh 安装预编译 wheel，不要在此直接 pip install

# 基础科学计算
numpy>=1.23.0,<2.0.0
pillow>=9.0.0
scipy>=1.9.0

# 数据可视化
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.5.0

# 目标检测辅助工具
pycocotools>=2.0.6
tqdm>=4.64.0

# XML 解析（VOC 标注）
lxml>=4.9.0

# Jittor（仅需代码正确，不实际运行）
# jittor>=1.3.0
